requirements.txt
aiofiles==25.1.0
aiohttp==3.13.3
beautifulsoup4
chromadb
langchain-text-splitters
livekit==1.0.25
livekit-agents==1.5.0
livekit-api==1.1.0
livekit-plugins-openai==1.5.0
numpy
openai==2.29.0
python-dotenv==1.2.2
requests==2.32.5
sentence-transformers
websockets==15.0.1
